% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{add_lead}
\alias{add_lead}
\title{Add leading zeros to a numeric vector
This function takes a numeric vector x and an integer width as inputs and returns a character vector of the same length as x with each element padded with leading zeros to match the specified width}
\usage{
add_lead(x, width = max(nchar(x)))
}
\arguments{
\item{x}{a numeric vector to be padded with leading zeros}

\item{width}{width an integer specifying the number of digits to pad each element of x with, default is the maximum number of characters in x}
}
\value{
a character vector of the same length as x with each element padded with leading zeros to match the specified width
}
\description{
Add leading zeros to a numeric vector
This function takes a numeric vector x and an integer width as inputs and returns a character vector of the same length as x with each element padded with leading zeros to match the specified width
}
\examples{
add_lead(c(1, 23, 456)) # returns c(“001”, “023”, “456”)
add_lead(c(1, 23, 456), width = 4) # returns c(“0001”, “0023”, “0456”)
}
