% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/begin-scofa.R
\name{begin_examcode_priorvers}
\alias{begin_examcode_priorvers}
\title{Starting an Analysis}
\usage{
begin_examcode_priorvers(...)
}
\arguments{
\item{, , , }{either empty for inserting the path to your exam folder, or if your exam folder is your working directoy, leave empty}
}
\value{
creates a directory set in your working directory and a list object with paths to important directories
}
\description{
Starting an Analysis
}
\examples{
begin_examcode_providers("NONEXAM")
}
