% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidden-reviewing-1.R
\name{cohort_count}
\alias{cohort_count}
\title{Count the number of 0s and 1s in the cohort vector and confirm with the user}
\usage{
cohort_count(facet1)
}
\arguments{
\item{facet1}{A data frame containing a \code{cohort} column.}
}
\description{
This function counts the number of 0s and 1s in the \code{cohort} vector of a data frame called \code{facet1}, and prompts the user to confirm whether the counts are correct using the \code{ui_yeah} function from the \code{usethis} package.
}
\keyword{internal}
