% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{ColNums_AllMissing}
\alias{ColNums_AllMissing}
\title{Find the column numbers of a data frame that have all missing values}
\usage{
ColNums_AllMissing(df)
}
\arguments{
\item{df}{df a data frame whose columns will be checked for missing values}
}
\value{
a vector of the column numbers that have all missing values in df
}
\description{
This function takes a data frame df as input and returns a vector of the column numbers that have all missing values (NA) in df
}
\examples{
df <- data.frame(x = c(1, 2, NA), y = c(NA, NA, NA), z = c(3, 4, 5))
}
