% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-ctrlfilewrite.R
\name{ctrl_write}
\alias{ctrl_write}
\title{Write and inspect control file}
\usage{
ctrl_write(
  pooled = TRUE,
  info_list = info_list,
  path = paths_list[["equated"]]
)
}
\arguments{
\item{pooled}{Logical. Pooled control file differs from unpooled. TRUE writes the pooled version. Unpooled version yet to be implemented.}

\item{info_list}{list object generated and updated throughout analysis}

\item{path}{This is the path for the control file. Its default is the equated subfolder of the un/pooled structure.}
}
\value{
A table is returned to the viewer with the control file open for inspection.
}
\description{
This writes an updated control file used by Calvin. Currently, it creates the pooled control file, but it will be updated to write all control files.
}
\examples{
ctrl_write(pooled = TRUE, info_list = info_list)
}
