% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidden-step0-info.R
\name{info_exam1}
\alias{info_exam1}
\title{Get early information about an exam using contextual information}
\usage{
info_exam1(exam_infolist = exam_info, revised_items = NULL)
}
\arguments{
\item{revised_items}{a list of revised items, which can be replaced with subsequent functions, in form list(ITEMID, ITEMID, ITEMID). Specifying the revised items in alternative list formats, vectors, or numeric with lead to issues at this stage of development.}
}
\value{
exam information in list form to be used in subsequent scoring functions
}
\description{
The function assumes a properly generated paths_list list and a list of revised items, which can be replaced with subsequent functions, in form list(ITEMID, ITEMID, ITEMID). Specifying the revised items in alternative list formats, vectors, or numeric with lead to issues at this stage of development.
}
\examples{
info_exam()
info_exam(list(984654, 243567))
}
