% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info-exam0.R
\name{info_exam_beta}
\alias{info_exam_beta}
\title{Initial info pull for present exam
\subsection{DOCUMENTATION IN PROGRESS}{
}}
\usage{
info_exam_beta(hobbes_in_wd = NULL, revised_items = NULL)
}
\arguments{
\item{hobbes_in_wd}{}
}
\description{
Initial info pull for present exam
\subsection{DOCUMENTATION IN PROGRESS}{
}
}
