% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-revised.R
\name{init_anchors}
\alias{init_anchors}
\title{Find revised items and store them into an info list}
\usage{
init_anchors(
  bench_csv = bench_csv,
  ed_xlsx = ed_xlsx,
  pooled = FALSE,
  write_anc = FALSE,
  info_list = info_list
)
}
\arguments{
\item{bench_csv}{Path to ITEMS_Total.csv winsteps output from an analysis considered benchmark. Use of {fs} package encouraged for generating paths.}

\item{ed_xlsx}{Path to ExamDeveloper Excel file with current item stats. Use of {fs} package encouraged for generating paths.}

\item{pooled}{Logical. This parameter may be used to place data, but it could be removed if info_list contains this information.}

\item{write_anc}{Logical. FALSE only stores the initial anchors, TRUE writes them}

\item{info_list}{name of the info_list generated by info_exam()}
}
\value{
info_list is returned, not assigned. As an optional side effect, the initial anchor file can be written.
}
\description{
Given 1) examdeveloper item output in xlsx format and 2) an item file with benchmark difficulties,
find the items that have been revised and optionally write an initial anchor file
}
\examples{
info_list <- init_anchors(bench_csv = fs::path_wd("data/ITEMS_Total.csv"),
                          ed_xlsx = fs::path_wd("data/ED.xlsx"),
                          write_anc = FALSE, info_list = info_list)
}
