% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_rev_move.R
\name{item_rev_move}
\alias{item_rev_move}
\title{Move Item Revisions}
\usage{
item_rev_move(ir_track_output = NULL)
}
\arguments{
\item{ir_track_output}{Item revision tracking output. Defaults to NULL.}
}
\value{
None.
}
\description{
This function moves item revisions by performing various operations on the provided item revision tracking output.
}
\examples{
item_rev_move()
}
