% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory-functions.R
\name{path_struc2}
\alias{path_struc2}
\title{Create a directory structure for pooled or unpooled data}
\usage{
path_struc2(exam_info = NULL, copy_hobbes = TRUE)
}
\arguments{
\item{exam_info}{= list generated from scofa::info_exam0();checks cohort n. create a pooled directory with subdirectories for data (scored responses, raw data, and keys) , item analysis, unequated and equated. If FALSE, create an unpooled directory with subdirectories for item analysis, unequated and equated.}

\item{copy_hobbes}{logical; if TRUE, copy the files from Hobbes Original into the equated subdirectory of the pooled or unpooled directory. If FALSE, do not copy the files.}
}
\value{
exam_info
}
\description{
This function creates a directory structure for pooled or unpooled data
and optionally copies the contents from Hobbes Original into the equated subdirectory.
}
\examples{
path_struc(pooled = TRUE, copy_hobbes = TRUE)# creates a pooled folder structure and copies files from Hobbes Original
path_struc(pooled = FALSE, copy_hobbes = FALSE) # creates an unpooled folder structure and does not copy files from Hobbes Original


}
