% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidden-reviewing-1.R
\name{read_items_csv}
\alias{read_items_csv}
\title{Read item data and compute P-value}
\usage{
read_items_csv(analysis_path = NULL, cohort_analysis = TRUE)
}
\arguments{
\item{analysis_path}{The path to the folder containing the analysis data. If NULL, defaults to the "pooled/equated" folder in the current working directory.}

\item{cohort_analysis}{Logical. If TRUE (default), reads in item data from the reanchored folder. If FALSE, reads in item data from the pooled/equated folder.}
}
\value{
A data frame containing the item data with an added "PVALUE" column.
}
\description{
This function reads in a CSV file with item data and computes the P-value for each item. The function can read in item data from either the pooled/equated folder or the reanchored folder, depending on the value of the cohort_analysis parameter.
}
\keyword{internal}
