% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidden-reviewing-1.R
\name{read_persons_csv}
\alias{read_persons_csv}
\title{Read person data and add cohort column}
\usage{
read_persons_csv(analysis_path = NULL)
}
\arguments{
\item{analysis_path}{The path to the folder containing the analysis data. If NULL, defaults to the "pooled/equated" folder in the current working directory.}
}
\value{
A data frame containing the person data with an added "cohort" column.
}
\description{
This function reads in a CSV file with candidate names and a CSV file with person data, and adds a "cohort" column to the person data based on whether each person's name is in the candidate name list.
}
\keyword{internal}
