% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{round.up}
\alias{round.up}
\title{Round up a numeric vector to a specified number of decimal places}
\usage{
\method{round}{up}(x, n)
}
\arguments{
\item{x}{a numeric vector to be rounded up}

\item{n}{an integer specifying the number of decimal places to round up to}
}
\value{
a numeric vector of the same length as x with each element rounded up to n decimal places
}
\description{
This function takes a numeric vector x and an integer n as inputs and returns a vector of the same length as x with each element rounded up to n decimal places
}
\examples{
round.up(c(1.23, 4.56, 7.89), 1) # returns c(1.3, 4.6, 7.9)
round.up(c(-1.23, -4.56, -7.89), 1) # returns c(-1.2, -4.5, -7.8)
}
