% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-runwinsteps.R
\name{run_winsteps}
\alias{run_winsteps}
\title{Run Winsteps cmd file}
\usage{
run_winsteps(path = paths_list[["equated"]])
}
\arguments{
\item{path}{Path to folder containing cmd file, defaults to the equated folder in the paths_list object. Encourage using {fs} package}
}
\value{
Prints the batch file command and returns a process identifier.
}
\description{
Run Winsteps cmd file
}
\examples{
run_winsteps() # defaults to equated subdirectory in either pooled or unpooled directory
run_winsteps(path = fs::path_wd("Hobbes Original"))
}
