% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{sd.p}
\alias{sd.p}
\title{returns the population standard deviation}
\usage{
sd.p(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
a number
}
\description{
returns the population standard deviation
}
\examples{
sd.p(c(6,5.4,7,6666,29.3))
}
