% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step1-pool.R
\name{write_respdat}
\alias{write_respdat}
\title{((FUNCTION BETA)) Write Response Data in _RESP.dat format (pooled) but later to include unpooled}
\usage{
write_respdat(info_list = info_list, datname = "_RESP.dat")
}
\arguments{
\item{info_list}{a list generated by by scofa with contextual information about the exam}

\item{datname}{the end of the RESP.dat filename, defaulting to {exam_code}_RESP.dat}
}
\value{
as a side effect, writes the _RESP.dat file to un/pooled directory and ....
}
\description{
((FUNCTION BETA)) Write Response Data in _RESP.dat format (pooled) but later to include unpooled
}
\examples{
write_respdat()
write_respdat(info_list = ilist, datname = "_RESP2.dat")
}
