% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{add_lead}
\alias{add_lead}
\title{Add leading zeros to a vector or list
This function pads a vector with leading zeros using \code{stringr::str_pad()}.}
\usage{
add_lead(x, width = NULL)
}
\arguments{
\item{x}{a list or vector of type string or numeric}

\item{width}{An integer. The desired width, in number of characters, of the output. If not provided, it is set to the maximum number of characters/digits in \code{x}.}
}
\value{
A vector of the same type as \code{x}, padded with leading zeros.
}
\description{
Add leading zeros to a vector or list
This function pads a vector with leading zeros using \code{stringr::str_pad()}.
}
\examples{
add_lead(c(1, 23, 456)) # returns c(“001”, “023”, “456”)
add_lead(c(1, 23, 456), width = 4) # returns c(“0001”, “0023”, “0456”)

}
\seealso{
\code{\link{str_pad}}
}
