% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidden-info-exam_beta.R
\name{info_exam_beta}
\alias{info_exam_beta}
\title{Initial Information Extraction for Present Exam}
\usage{
info_exam_beta(create_exam_dir = TRUE)
}
\arguments{
\item{hobbes_in_wd}{A string specifying the directory in the working directory where the hobbes files are stored. Default is NULL.}
}
\value{
A list named 'exam_info' with the following elements:
\itemize{
\item 'exam_code': The name of the current working directory, which should be the exam code.
\item 'cohort_n': The number of rows in the pre-scored data, which should be the cohort size.
\item 'nitms_form': The number of columns minus 1 in the pre-scored data, which should be the number of items per form.
\item 'n_opts': The number of unique response options in the domain key.
\item 'i_opts': Number of distinct options.
\item 'codes': A string of response option codes. If there are more than 5, a warning is printed.
}
}
\description{
This function extracts initial information from the present exam data.
The data should be in a directory named 'Hobbes Original' within the working directory.
The information extracted includes the exam code, cohort size, the number of items per form,
the number of response options, and the response option codes.
}
\examples{
\dontrun{
exam_info = info_exam_beta()
}
}
